# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NaikPete' API",
        "description": "API documentation for My Application",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/transaksi": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Get all transactions for the authenticated user",
                "operationId": "926153a57fbd396e3cb14c30b99ca73b",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User tidak terautentikasi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No transactions found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tidak ada transaksi untuk user ini"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/driver/transaksi": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Get all transactions for the authenticated driver",
                "operationId": "fb454993cc1e6bd974c478ba06c75696",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved driver transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Driver only can access this",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hanya driver yang dapat melihat transaksi ini"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No transactions found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tidak ada transaksi untuk driver ini"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transaksi/filter/date/{date}": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Filter transactions by date",
                "operationId": "d1eca20cdc785eb7a020c1433d340cf9",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "The date to filter transactions",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully filtered transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transaksi/filter/month/{month}/year/{year}": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Filter transactions by month and year",
                "operationId": "4f06e4b8f7dfbeb43148b96930a440c6",
                "parameters": [
                    {
                        "name": "month",
                        "in": "path",
                        "description": "Month to filter",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "year",
                        "in": "path",
                        "description": "Year to filter",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully filtered transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transaksi/filter/year/{year}": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Filter transactions by year",
                "operationId": "a415c993e2f79a2e4d17dcf352a153d8",
                "parameters": [
                    {
                        "name": "year",
                        "in": "path",
                        "description": "Year to filter",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully filtered transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transaksi/filter/date/{date}/month/{month}/year/{year}": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Filter transactions by specific date, month, and year",
                "operationId": "d4224be5b111a6662020d678c51466f4",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "The date to filter",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "month",
                        "in": "path",
                        "description": "Month to filter",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "year",
                        "in": "path",
                        "description": "Year to filter",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully filtered transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Driver": {
                "required": [
                    "id_driver",
                    "nama",
                    "email",
                    "password",
                    "alamat",
                    "tgl_lahir",
                    "no_hp",
                    "gender",
                    "id_kendaraan"
                ],
                "properties": {
                    "id_driver": {
                        "description": "ID of the driver",
                        "type": "integer"
                    },
                    "nama": {
                        "description": "Name of the driver",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email of the driver",
                        "type": "string"
                    },
                    "password": {
                        "description": "Password of the driver",
                        "type": "string"
                    },
                    "alamat": {
                        "description": "Address of the driver",
                        "type": "string"
                    },
                    "tgl_lahir": {
                        "description": "Date of birth of the driver",
                        "type": "string",
                        "format": "date"
                    },
                    "no_hp": {
                        "description": "Phone number of the driver",
                        "type": "string"
                    },
                    "gender": {
                        "description": "Gender of the driver",
                        "type": "string"
                    },
                    "id_kendaraan": {
                        "description": "ID of the vehicle assigned to the driver",
                        "type": "integer"
                    },
                    "kendaraan": {
                        "$ref": "#/components/schemas/Kendaraan"
                    }
                },
                "type": "object"
            },
            "Halte": {
                "required": [
                    "id",
                    "nama_halte",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the halte",
                        "type": "integer"
                    },
                    "nama_halte": {
                        "description": "Name of the halte",
                        "type": "string"
                    },
                    "latitude": {
                        "description": "Latitude of the halte",
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "description": "Longitude of the halte",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "Kendaraan": {
                "required": [
                    "id",
                    "trayek_id",
                    "no_plat",
                    "tahun_kendaraan",
                    "is_tersedia",
                    "koordinat"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the vehicle",
                        "type": "integer"
                    },
                    "trayek_id": {
                        "description": "ID of the route (trayek) the vehicle is assigned to",
                        "type": "integer"
                    },
                    "no_plat": {
                        "description": "License plate number of the vehicle",
                        "type": "string"
                    },
                    "tahun_kendaraan": {
                        "description": "Year of the vehicle",
                        "type": "integer"
                    },
                    "is_tersedia": {
                        "description": "Availability of the vehicle (true if available, false if not)",
                        "type": "boolean"
                    },
                    "koordinat": {
                        "description": "Coordinates of the vehicle (e.g., latitude, longitude)",
                        "type": "string"
                    },
                    "trayek": {
                        "$ref": "#/components/schemas/Trayek"
                    },
                    "schedules": {
                        "$ref": "#/components/schemas/Schedule"
                    },
                    "drivers": {
                        "$ref": "#/components/schemas/Driver"
                    }
                },
                "type": "object"
            },
            "Schedule": {
                "required": [
                    "id",
                    "id_kendaraan",
                    "waktu_berangkat",
                    "waktu_tiba"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the schedule",
                        "type": "integer"
                    },
                    "id_kendaraan": {
                        "description": "ID of the kendaraan (vehicle) associated with the schedule",
                        "type": "integer"
                    },
                    "waktu_berangkat": {
                        "description": "Departure time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "waktu_tiba": {
                        "description": "Arrival time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "kendaraan": {
                        "$ref": "#/components/schemas/Kendaraan"
                    },
                    "trayeks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Trayek"
                        }
                    }
                },
                "type": "object"
            },
...
            "Transaction": {
                "required": [
                    "id",
                    "id_user",
                    "id_driver",
                    "id_trayek",
                    "jenis_tiket",
                    "tgl_transaksi",
                    "jumlah_tiket",
                    "payment_method",
                    "payment_amount",
                    "payment_status"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the transaction",
                        "type": "integer"
                    },
                    "id_user": {
                        "description": "ID of the user who made the transaction",
                        "type": "integer"
                    },
                    "id_driver": {
                        "description": "ID of the driver involved in the transaction",
                        "type": "integer"
                    },
                    "id_trayek": {
                        "description": "ID of the route (trayek) associated with the transaction",
                        "type": "integer"
                    },
                    "jenis_tiket": {
                        "description": "Type of ticket (e.g., regular, VIP)",
                        "type": "string"
                    },
                    "tgl_transaksi": {
                        "description": "Transaction date and time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "jumlah_tiket": {
                        "description": "Number of tickets purchased",
                        "type": "integer"
                    },
                    "payment_method": {
                        "description": "Payment method used (e.g., cash, credit card)",
                        "type": "string"
                    },
                    "payment_amount": {
                        "description": "Amount paid for the transaction",
                        "type": "number",
                        "format": "float"
                    },
                    "payment_status": {
                        "description": "Status of the payment (e.g., pending, completed)",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "driver": {
                        "$ref": "#/components/schemas/Driver"
                    },
                    "trayek": {
                        "$ref": "#/components/schemas/Trayek"
                    }
                },
                "type": "object"
            },
            "Trayek": {
                "required": [
                    "id",
                    "kode_trayek",
                    "urutan_halte",
                    "id_schedule"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the trayek",
                        "type": "integer"
                    },
                    "kode_trayek": {
                        "description": "Trayek code",
                        "type": "string"
                    },
                    "urutan_halte": {
                        "description": "Order of halts",
                        "type": "integer"
                    },
                    "id_schedule": {
                        "description": "Associated schedule ID",
                        "type": "integer"
                    },
                    "schedule": {
                        "$ref": "#/components/schemas/Schedule"
                    },
                    "haltes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Halte"
                        }
                    },
                    "kendaraan": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Kendaraan"
                        }
                    },
                    "transaksi": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Transaction"
                        }
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "no_hp": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Merdeka No. 123"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "laki-laki",
                            "perempuan"
                        ],
                        "example": "laki-laki"
                    },
                    "tgl_lahir": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "customer",
                            "driver"
                        ],
                        "example": "customer"
                    },
                    "foto_profil": {
                        "description": "Optional profile photo",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            }
        }
    },
...
}